Refactor question flow
